Make conversation deletion more resilient
