+++ front/lib/file_storage/index.ts
@@ -107,6 +107,12 @@ export class FileStorage {
     return this.bucket.file(filename);
   }
 
+  async getFiles({ prefix }: { prefix?: string } = {}) {
+    const [files] = await this.bucket.getFiles({ prefix });
+
+    return files;
+  }
+
   get name() {
     return this.bucket.name;
   }
@@ -151,3 +157,6 @@ export const getPublicUploadBucket = (options?: FileStorageOptions) =>
 
 export const getUpsertQueueBucket = (options?: FileStorageOptions) =>
   getBucketInstance(config.getGcsUpsertQueueBucket(), options);
+
+export const getDustDataSourcesBucket = (options?: FileStorageOptions) =>
+  getBucketInstance(config.getDustDataSourcesBucket(), options);
